Add tests for transport encodings
